// Bullhorn-Attachments-UI-Code/src/app/shared/models/review/employment-data.model.ts
export class Employment {
    fusionPersonId: string | null | undefined;
    fusionEmployeeId: string | null | undefined;
    name: string | null | undefined;
    employmentStatus: string | null | undefined;
    personType: string | null | undefined;
    fusionSupplierName: string | null | undefined;
    firstName: string | null | undefined;
    middleName: string | null | undefined;
    lastName: string | null | undefined;
    preferredName: string | null | undefined;
    preferredLanguage: string | null | undefined;
    gender: string | null | undefined;
    birthDate: string | null | undefined;
    taxId: string | null | undefined;
    partTimeOrFullTime: string | null | undefined;
    overallTenure: string | null | undefined;
    currentTenure: string | null | undefined;
    startDateEmployment: string | null | undefined;
    originalStartDate: string | null | undefined;
    mostRecentTerminationDate: string | null | undefined;
    endDateEmployment: string | null | undefined;
    payFrequency: string | null | undefined;
    payWeekendingDay: string | null | undefined;
    payCycle: string | null | undefined;
    payGroup: string | null | undefined;
    payRateType: string | null | undefined;
    payCurrency: string | null | undefined;
    overtimePayRate: string | null | undefined;
    doubleTimePayRate: string | null | undefined;
    legalEmployer: string | null | undefined;
    reportingBusinessUnit: string | null | undefined;
    brandOffering: string | null | undefined;
    lineOfBusiness: string | null | undefined;
    location: string | null | undefined;
    jobTitle: string | null | undefined;
    jobClass: string | null | undefined;
    exemptionStatus: string | null | undefined;
    exemptionReason: string | null | undefined;
    homeAddress: string | null | undefined;
    primaryEmail: string | null | undefined;
    businessEmail: string | null | undefined;
    mobilePhone: string | null | undefined;
    homePhone: string | null | undefined;
    preferredContactMethod: string | null | undefined;
    addressType: string | null | undefined;
    streetAddress1: string | null | undefined;
    streetAddress2: string | null | undefined;
    streetAddress3: string | null | undefined;
    city: string | null | undefined;
    county: string | null | undefined;
    state: string | null | undefined;
    country: string | null | undefined;
    zip: string | null | undefined;
    overwriteReason: string | null | undefined;
    googleValidationStatus: string | null | undefined;
    immigrationStatus: string | null | undefined;
    visaType: string | null | undefined;
    visaStartDate: string | null | undefined;
    visaEndDate: string | null | undefined;
    visaNumber: string | null | undefined;
    supplierPayGroup: string | null | undefined;
    usW2ElectronicConsent: string | null | undefined;
    federalGovernmentWorker: string | null | undefined;
    payrollId: string | null | undefined;
    sourceSystem: string | null | undefined;
    ptoPlan: string | null | undefined;
    balance: string | null | undefined;
    asofPeriodEnding: string | null | undefined;
    jurisdictionBalance: string | null | undefined;
    medicalBenefitsEligibility: string | null | undefined;
    holidayRule: string | null | undefined;
    personUnqId: string | null | undefined;
    legacyEmployeeId:string | null | undefined;
    legacyAssignmentId:string | null | undefined;
    placementId:string | null | undefined;
    candidateId:string | null | undefined;
    immigrationNotes:string | null | undefined;
    financeLocation:string | null | undefined;    
    constructor(data: any) {
        Object.assign(this, data);
    }
}